// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/programs-model.ts
export class ProgramModel {
    isPagingRequired: boolean
    programId: string;
    dataSetId: string;
    dataSourceId: string;
    id: string
    isArchived: boolean;
    programName: string;
    programData: string;
    location: string;
    contractType: string;
    formData: any;
    userIds: string[];
    template : string;
    imageUrl: string;
    isNewRecord: any;
    programUniqueId: any;
}